+++ .gitignore
@@ -1,1 +1,1 @@
-basic_animation.mp4
+dest/basic_animation.mp4
