--- kotani_branch_test.txt
@@ -1,1 +1,0 @@
-ブランチが正しく作成できたかのテストです。
